store user_id as a string in session, instead of a Long
When a snowflake is stored in a session as a Long, it loses some information, including its type, apparently. This makes it impossible (?) to detect Longs and convert them to strings later. I don't think there's really any benefit at this point to passing snowflakes around as Longs, so, we'll just kinda phase them out as they appear, I guess.